# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.Description:
    properties:
      caption:
        type: string
      id:
        type: integer
      text:
        type: string
    type: object
...
  domain.Faq:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    type: object
...
  domain.GetAllDescCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Description'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
  domain.GetAllFaqsCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Faq'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
...
  domain.PaginationPage:
    properties:
      count:
        type: integer
      page:
        type: integer
      pages:
        type: integer
    type: object
...
  v1.Description:
    properties:
      caption:
        type: string
      text:
        type: string
    type: object
  v1.Faq:
    properties:
      answer:
        type: string
      question:
        type: string
    type: object
...
  v1.idResponse:
    properties:
      id: {}
    type: object
  v1.okResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.response:
    properties:
      detail:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: API Server for  A Hundred to Five
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: A Hundred to Five
  version: "2.0"
paths:
...
  /description:
    get:
      consumes:
      - application/json
      description: get all description
      operationId: get-all-description
      parameters:
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllDescCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - about
    post:
      consumes:
      - application/json
      parameters:
      - description: description
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.Description'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - about
  /description/{id}:
    delete:
      consumes:
      - application/json
      description: delete description
      parameters:
      - description: description id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - about
    get:
      consumes:
      - application/json
      description: get description by id
      operationId: get-description-by-id
      parameters:
      - description: description id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Description'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - about
    put:
      consumes:
      - application/json
      description: update  description
      parameters:
      - description: description id
        in: path
        name: id
        required: true
        type: string
      - description: description
        in: body
        name: input
        schema:
          $ref: '#/definitions/v1.Description'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - about
  /faq:
    get:
      consumes:
      - application/json
      description: get all faqs
      operationId: get-all-faqs
      parameters:
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllFaqsCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - faq
    post:
      consumes:
      - application/json
      parameters:
      - description: faq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.Faq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - faq
  /faq/{id}:
    delete:
      consumes:
      - application/json
      description: delete faq
      parameters:
      - description: faq id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - faq
    get:
      consumes:
      - application/json
      description: get faq by id
      operationId: get-faq-by-id
      parameters:
      - description: faq id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Faq'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - faq
    put:
      consumes:
      - application/json
      description: update  faq
      parameters:
      - description: faq id
        in: path
        name: id
        required: true
        type: string
      - description: faq
        in: body
        name: input
        schema:
          $ref: '#/definitions/v1.Faq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - faq
...
securityDefinitions:
  User_Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"